todo_backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
PyJWT==2.8.0
python-dotenv==1.0.1